item shop is real and stuff i forgor

+++ src/commands/work.ts
@@ -11,8 +11,8 @@ export default {
         var { getUser } = getHotReloadable().eco
         var u = await getUser(msg.author)
         var mul = getMul(u)
-        var points = Math.floor(Math.random() * 25 * mul)
-        var gold = Math.floor(Math.random() * 3 * mul)
+        var points = BigInt(Math.floor(Math.random() * 25)) * mul
+        var gold = BigInt(Math.floor(Math.random() * 3)) * mul
         u.money.points += points
         u.money.gold += gold
         await msg.reply(`Earned ${moneyFormat(points)} and ${moneyFormat(gold, "gold")}`)
